🎨 [Refactor] : nextAuth util 함수 분리
