+++ src/app/auth/profile/page.tsx
@@ -1,10 +1,10 @@
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
 import { getMyPage } from '@/utils/fetch';
+import { authOptions } from '@/utils/next-auth';
 import { getServerSession } from 'next-auth';
 import UserDetailsEditor from './_components/UserDetailsEditor';
 
 // fetch 데이터가 공통적으로 사용해서 병렬라우팅 불가
-export default async function ProfilePage({ a }: { a: number }) {
+export default async function ProfilePage() {
   const user = await getServerSession(authOptions);
   const data = await getMyPage({
     memberId: user?.user.memberId!,
